backend/upload-scan-result-api/aeternalis_api/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
filelock==3.18.0
fsspec==2025.5.1
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
urllib3==2.4.0
